+++ tldr_app/tldr-dekstop/tldr_cpp/src/lib_tldr/db/sqlite_database.h
@@ -12,7 +12,7 @@ namespace tldr {
         ~SQLiteDatabase() override;
 
         bool initialize() override;
-        int64_t saveEmbeddings(const std::vector<std::string_view> &chunks, const json &embeddings_response, const std::vector<size_t> &embedding_hashes = {}) override;
+        int64_t saveEmbeddings(const std::vector<std::string_view> &chunks, const json &embeddings_response, const std::vector<uint64_t> &embedding_hashes = {}) override;
         bool getEmbeddings(int64_t id, std::vector<std::string> &chunks, json &embeddings) override;
         std::vector<std::pair<std::string, float>> searchSimilarVectors(const std::vector<float>& query_vector, int k = 5) override;
         std::map<uint64_t, std::string> getChunksByHashes(const std::vector<uint64_t>& hashes) override;
