+++ tldr_app/tldr-dekstop/tldr_cpp/src/lib_tldr/lib_tldr.h
@@ -50,13 +50,13 @@ std::string extractTextFromPDF(const std::string &filename);
 std::vector<std::string>
 splitTextIntoChunks(const std::string &text, size_t max_chunk_size = 2000, size_t overlap = 20);
 bool initializeDatabase();
-int64_t saveEmbeddingsToDb(const std::vector<std::string_view> &chunks, const std::vector<std::vector<float>> &embeddings, const std::vector<size_t> &embeddings_hash = {});
+int64_t saveEmbeddingsToDb(const std::vector<std::string_view> &chunks, const std::vector<std::vector<float>> &embeddings, const std::vector<uint64_t> &embeddings_hash = {});
 std::string sendEmbeddingsRequest(const json &request, const std::string& url);
 json parseEmbeddingsResponse(const std::string &response_data);
-int saveEmbeddingsThreadSafe(const std::vector<std::string> &batch, const std::vector<std::vector<float>> &batch_embeddings, const std::vector<size_t> &embeddings_hash);
+int saveEmbeddingsThreadSafe(const std::vector<std::string> &batch, const std::vector<std::vector<float>> &batch_embeddings, const std::vector<uint64_t> &embeddings_hash);
 void processChunkBatch(const std::vector<std::string> &batch, size_t batch_num, size_t total_batches, int &result_id);
 // Returns gathered embeddings and their hashes
-std::pair<std::vector<std::vector<float>>, std::vector<size_t>> 
+std::pair<std::vector<std::vector<float>>, std::vector<uint64_t>> 
 obtainEmbeddings(const std::vector<std::string> &chunks, size_t batch_size = 2, size_t num_threads = 2);
 
 // Include vector dump functionality
